+++ lib/image_util.rb
@@ -6,12 +6,14 @@ module ImageUtil
   class Error < StandardError; end
   # Your code goes here...
 
+  autoload :BitmapFont, "image_util/bitmap_font"
   autoload :Color, "image_util/color"
   autoload :Image, "image_util/image"
   autoload :Util, "image_util/util"
   autoload :Codec, "image_util/codec"
   autoload :Magic, "image_util/magic"
   autoload :Filter, "image_util/filter"
+  autoload :Generator, "image_util/generator"
   autoload :Statistic, "image_util/statistic"
   autoload :Terminal, "image_util/terminal"
   autoload :View, "image_util/view"