+++ lib/image_util/generator.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ImageUtil
+  module Generator
+    autoload :BitmapText, "image_util/generator/bitmap_text"
+  end
+end